refactor: 딜러와 플레이어 중복 기능 상속으로 리팩토링
