+++ src/main/java/domain/participant/Player.java
@@ -14,18 +14,10 @@ public class Player extends Participant {
     }
 
     public boolean checkCardsCondition() {
-        return getTotalValue() <= LIMIT_TAKE_CARD_VALUE;
+        return getMaxSum() <= LIMIT_TAKE_CARD_VALUE;
     }
 
     public List<Card> getCards() {
         return Collections.unmodifiableList(cards);
-    }
-
-    public int getTotalValue() {
-        int totalValue = 0;
-        for (Card card : cards) {
-            totalValue += card.getValue();
-        }
-        return totalValue;
     }
 }
